% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a vcf file}
\usage{
read_vcf(vcfFile)
}
\arguments{
\item{vcfFile}{Path to vcf file}
}
\value{
A list with elements header, vm, vi
}
\description{
Read a vcf file
}
\examples{
\dontrun{
vcf<-read_vcf("inst/extdata/trio.2010_06.ychr.sites.vcf.gz")
}
}
